+++ dsmr_mqtt/admin.py
@@ -1,4 +1,6 @@
 from django.contrib import admin
+from django.forms import TextInput
+from django.db import models
 from django.utils.translation import gettext_lazy as _
 from solo.admin import SingletonModelAdmin
 
@@ -53,16 +55,18 @@ class RawTelegramMQTTSettingsAdmin(SingletonModelAdmin):
 
 @admin.register(telegram.JSONTelegramMQTTSettings)
 class JSONTelegramMQTTSettingsAdmin(SingletonModelAdmin):
+    formfield_overrides = {
+        models.CharField: {"widget": TextInput(attrs={"size": "72"})},
+    }
     fieldsets = (
         (
             None,
             {
                 "fields": ["enabled", "topic", "formatting", "use_local_timezone"],
-                "description": _(
-                    "Triggered by any method of reading insertion (datalogger or API). "
-                    "Allows you to send newly created readings to the MQTT broker, as a JSON message. You can alter "
-                    "the field names used in the JSON message. Removing lines will remove fields from the message as "
-                    "well. "
+                "description": _("Triggered by any method of reading insertion (datalogger or API).") + " " +
+                _(
+                    "Allows you to send this data to the MQTT broker, as a JSON message. You can alter the field names "
+                    "used in the JSON message. Removing lines will remove fields from the message as well. "
                     """Default value:
                     <pre>
                     {}
@@ -87,11 +91,10 @@ class SplitTopicTelegramMQTTSettingsAdmin(SingletonModelAdmin):
             None,
             {
                 "fields": ["enabled", "formatting", "use_local_timezone"],
-                "description": _(
-                    "Triggered by any method of reading insertion (datalogger or API). "
-                    "Allows you to send newly created readings to the MQTT broker, splitted per field. You can "
-                    "designate each field name to a different topic. Removing lines will prevent those fields from "
-                    "being broadcast as well. "
+                "description": _("Triggered by any method of reading insertion (datalogger or API).") + " " +
+                _(
+                    "Allows you to send this data to the MQTT broker, split per-field. You can designate each field "
+                    "name to a different topic. Removing lines will prevent those fields from being broadcast as well. "
                     """Default value:
                     <pre>
                     {}
@@ -111,15 +114,18 @@ class SplitTopicTelegramMQTTSettingsAdmin(SingletonModelAdmin):
 
 @admin.register(day_totals.JSONDayTotalsMQTTSettings)
 class JSONDayTotalsMQTTSettingsAdmin(SingletonModelAdmin):
+    formfield_overrides = {
+        models.CharField: {"widget": TextInput(attrs={"size": "72"})},
+    }
     fieldsets = (
         (
             None,
             {
                 "fields": ["enabled", "topic", "formatting"],
-                "description": _(
-                    "Triggered by any method of reading insertion (datalogger or API). "
-                    "Send the current day totals to the broker. You can alter the the field names used in the JSON "
-                    "message. Removing lines will remove fields from the message as well. "
+                "description": _("Triggered by any method of reading insertion (datalogger or API).") + " " +
+                _(
+                    "Allows you to send this data to the MQTT broker, as a JSON message. You can alter the field names "
+                    "used in the JSON message. Removing lines will remove fields from the message as well. "
                     """Default value:
                     <pre>
                     {}
@@ -144,11 +150,10 @@ class SplitTopicDayTotalsMQTTSettingsAdmin(SingletonModelAdmin):
             None,
             {
                 "fields": ["enabled", "formatting"],
-                "description": _(
-                    "Triggered by any method of reading insertion (datalogger or API). "
-                    "Allows you to send day totals (dashboard) to the MQTT broker, splitted per field. You can "
-                    "designate each field name to a different topic. Removing lines will prevent those fields from "
-                    "being broadcast as well. "
+                "description": _("Triggered by any method of reading insertion (datalogger or API).") + " " +
+                _(
+                    "Allows you to send this data to the MQTT broker, split per-field. You can designate each field "
+                    "name to a different topic. Removing lines will prevent those fields from being broadcast as well. "
                     """Default value:
                     <pre>
                     {}
@@ -168,14 +173,18 @@ class SplitTopicDayTotalsMQTTSettingsAdmin(SingletonModelAdmin):
 
 @admin.register(period_totals.JSONCurrentPeriodTotalsMQTTSettings)
 class JSONPeriodTotalsUpdateMQTTSettingsAdmin(SingletonModelAdmin):
+    formfield_overrides = {
+        models.CharField: {"widget": TextInput(attrs={"size": "72"})},
+    }
     fieldsets = (
         (
             None,
             {
                 "fields": ["enabled", "topic", "formatting"],
-                "description": _(
-                    "Contains the totals for the current month and current year. "
-                    "Payload is a customizable JSON structure. "
+                "description": _("Contains the totals for the current month and current year.") + " " +
+                _(
+                    "Allows you to send this data to the MQTT broker, as a JSON message. You can alter the field names "
+                    "used in the JSON message. Removing lines will remove fields from the message as well. "
                     """Default value:
                     <pre>
                     {}
@@ -200,9 +209,10 @@ class SplitTopicPeriodTotalsUpdateMQTTSettingsAdmin(SingletonModelAdmin):
             None,
             {
                 "fields": ["enabled", "formatting"],
-                "description": _(
-                    "Contains the totals for the current month and current year. "
-                    "Payload is sent on a per field per topic basis. "
+                "description": _("Contains the totals for the current month and current year.") + " " +
+                _(
+                    "Allows you to send this data to the MQTT broker, split per-field. You can designate each field "
+                    "name to a different topic. Removing lines will prevent those fields from being broadcast as well. "
                     """Default value:
                     <pre>
                     {}
@@ -227,11 +237,10 @@ class SplitTopicMeterStatisticsMQTTSettingsAdmin(SingletonModelAdmin):
             None,
             {
                 "fields": ["enabled", "formatting"],
-                "description": _(
-                    "Triggered by any method of reading insertion (datalogger or API). "
-                    "Allows you to send meter statistics to the MQTT broker, splitted per field. You can "
-                    "designate each field name to a different topic. Removing lines will prevent those fields from "
-                    "being broadcast as well. "
+                "description": _("Triggered by any method of reading insertion (datalogger or API).") + " " +
+                _(
+                    "Allows you to send this data to the MQTT broker, split per-field. You can designate each field "
+                    "name to a different topic. Removing lines will prevent those fields from being broadcast as well. "
                     """Default value:
                     <pre>
                     {}
@@ -251,16 +260,18 @@ class SplitTopicMeterStatisticsMQTTSettingsAdmin(SingletonModelAdmin):
 
 @admin.register(consumption.JSONGasConsumptionMQTTSettings)
 class JSONGasConsumptionMQTTSettingsAdmin(SingletonModelAdmin):
+    formfield_overrides = {
+        models.CharField: {"widget": TextInput(attrs={"size": "72"})},
+    }
     fieldsets = (
         (
             None,
             {
                 "fields": ["enabled", "topic", "formatting"],
-                "description": _(
-                    "Triggered when a different gas reading is processed. "
-                    "Allows you to send gas consumption to the MQTT broker, as a JSON message. You can alter "
-                    "the field names used in the JSON message. Removing lines will remove fields from the message as "
-                    "well. "
+                "description": _("Triggered when a different gas reading is processed.") + " " +
+                _(
+                    "Allows you to send this data to the MQTT broker, as a JSON message. You can alter the field names "
+                    "used in the JSON message. Removing lines will remove fields from the message as well. "
                     """Default value:
                     <pre>
                     {}
@@ -285,11 +296,10 @@ class SplitTopicGasConsumptionMQTTSettingsAdmin(SingletonModelAdmin):
             None,
             {
                 "fields": ["enabled", "formatting"],
-                "description": _(
-                    "Triggered when a different gas reading is processed. "
-                    "Allows you to send gas consumption to the MQTT broker, splitted per field. You can "
-                    "designate each field name to a different topic. Removing lines will prevent those fields from "
-                    "being broadcast as well. "
+                "description": _("Triggered when a different gas reading is processed.") + " " +
+                _(
+                    "Allows you to send this data to the MQTT broker, split per-field. You can designate each field "
+                    "name to a different topic. Removing lines will prevent those fields from being broadcast as well. "
                     """Default value:
                     <pre>
                     {}
@@ -298,6 +308,67 @@ class SplitTopicGasConsumptionMQTTSettingsAdmin(SingletonModelAdmin):
                         [
                             x.default
                             for x in consumption.SplitTopicGasConsumptionMQTTSettings._meta.get_fields()
+                            if x.name == "formatting"
+                        ][0]
+                    )
+                ),
+            },
+        ),
+    )
+
+
+@admin.register(consumption.JSONQuarterHourPeakElectricityConsumptionMQTTSettings)
+class JSONQuarterHourPeakElectricityConsumptionMQTTSettingsAdmin(SingletonModelAdmin):
+    formfield_overrides = {
+        models.CharField: {"widget": TextInput(attrs={"size": "72"})},
+    }
+    fieldsets = (
+        (
+            None,
+            {
+                "fields": ["enabled", "topic", "formatting"],
+                "description": _("Triggered when a quarter-hour peak consumption is calculated.") + " " +
+                _(
+                    "Allows you to send this data to the MQTT broker, as a JSON message. You can alter the field names "
+                    "used in the JSON message. Removing lines will remove fields from the message as well. "
+                    """Default value:
+                    <pre>
+                    {}
+                    </pre>
+                    """.format(
+                        [
+                            x.default
+                            for x in consumption.JSONQuarterHourPeakElectricityConsumptionMQTTSettings._meta.get_fields()
+                            if x.name == "formatting"
+                        ][0]
+                    )
+                ),
+            },
+        ),
+    )
+
+
+@admin.register(consumption.SplitTopicQuarterHourPeakElectricityConsumptionMQTTSettings)
+class SplitTopicQuarterHourPeakElectricityConsumptionMQTTSettingsAdmin(
+    SingletonModelAdmin
+):
+    fieldsets = (
+        (
+            None,
+            {
+                "fields": ["enabled", "formatting"],
+                "description": _("Triggered when a quarter-hour peak consumption is calculated.") + " " +
+                _(
+                    "Allows you to send this data to the MQTT broker, split per-field. You can designate each field "
+                    "name to a different topic. Removing lines will prevent those fields from being broadcast as well. "
+                    """Default value:
+                    <pre>
+                    {}
+                    </pre>
+                    """.format(
+                        [
+                            x.default
+                            for x in consumption.SplitTopicQuarterHourPeakElectricityConsumptionMQTTSettings._meta.get_fields()
                             if x.name == "formatting"
                         ][0]
                     )